8-content-orchestration/content_orchestration/requirements.txt
google-adk==1.16.0
google-generativeai>=0.8.5
python-dotenv>=1.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.6

